Refactor getGridCell to take integer parameters and add some props

+++ gravity-2d/Game.hpp
@@ -19,7 +19,7 @@ public:
 
 	bool running() { return isRunning; }
 
-	GridCell* getGridCell(double x, double y);
+	GridCell* getGridCell(int x, int y);
 
 private:
 	bool isRunning;
